Implement new ai test generation panel
