+++ frontend/src/pages/RequestorPage/ai/index.ts
@@ -1,3 +1,4 @@
 export { useAi, FRIENDLY, HOSTILE, type AiTestingPersona } from "./ai";
 export { useSummarizeError } from "./summarize-traces";
-export { AiTestGeneration } from "./AiTestGeneration";
+export { AiTestGeneration } from "./AiTestGenerationDrawer";
+export { AiTestGenerationPanel } from "./AiTestGenerationPanel";
